+++ test/unit/mysql_db_admin/code_coverage.sh
@@ -9,7 +9,20 @@ coverage erase
 echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/help_message.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/analyze.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/check.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/checksum.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/detect_dbs.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/optimize.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/process_request.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_check.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_checksum.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_optimize.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_analyze.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/main.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_program.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/status.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/setup_mail.py
 
 echo ""
 echo "Producing code coverage report"