+++ test/unit/mysql_db_admin/run_check.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python
 # Classification (U)
 
-"""Program:  run_analyze.py
+"""Program:  run_check.py
 
-    Description:  Unit testing of run_analyze in mysql_db_admin.py.
+    Description:  Unit testing of run_check in mysql_db_admin.py.
 
     Usage:
-        test/unit/mysql_db_admin/run_analyze.py
+        test/unit/mysql_db_admin/run_check.py
 
     Arguments:
 
@@ -75,7 +75,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
-        test_run_analyze -> Test run_analyze function.
+        test_run_check -> Test run_check function.
 
     """
 
@@ -91,27 +91,27 @@ class UnitTest(unittest.TestCase):
 
         self.server = Server()
 
-        self.analyze_tables = [{"Msg_type": "Type", "Msg_text": "Message"},
-                               {"Msg_type": "Type2", "Msg_text": "Message2"}]
+        self.check_tbl = [{"Msg_type": "Type", "Msg_text": "Message"},
+                          {"Msg_type": "Type2", "Msg_text": "Message2"}]
 
     @mock.patch("mysql_db_admin.gen_libs.prt_msg")
-    @mock.patch("mysql_db_admin.mysql_libs.analyze_tbl")
-    def test_run_analyze(self, mock_analyze, mock_prt):
+    @mock.patch("mysql_db_admin.mysql_libs.check_tbl")
+    def test_run_check(self, mock_check, mock_prt):
 
-        """Function:  test_run_analyze
+        """Function:  test_run_check
 
-        Description:  Test run_analyze function.
+        Description:  Test run_check function.
 
         Arguments:
 
         """
 
-        mock_analyze.return_value = self.analyze_tables
+        mock_check.return_value = self.check_tbl
         mock_prt.return_value = True
 
         with gen_libs.no_std_out():
-            self.assertFalse(mysql_db_admin.run_analyze(self.server, "db",
-                                                        "tbl"))
+            self.assertFalse(mysql_db_admin.run_check(self.server, "db",
+                                                      "tbl"))
 
 
 if __name__ == "__main__":