+++ test/unit/mysql_db_admin/run_checksum.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python
 # Classification (U)
 
-"""Program:  run_analyze.py
+"""Program:  run_checksum.py
 
-    Description:  Unit testing of run_analyze in mysql_db_admin.py.
+    Description:  Unit testing of run_checksum in mysql_db_admin.py.
 
     Usage:
-        test/unit/mysql_db_admin/run_analyze.py
+        test/unit/mysql_db_admin/run_checksum.py
 
     Arguments:
 
@@ -75,7 +75,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
-        test_run_analyze -> Test run_analyze function.
+        test_run_checksum -> Test run_checksum function.
 
     """
 
@@ -91,27 +91,24 @@ class UnitTest(unittest.TestCase):
 
         self.server = Server()
 
-        self.analyze_tables = [{"Msg_type": "Type", "Msg_text": "Message"},
-                               {"Msg_type": "Type2", "Msg_text": "Message2"}]
+        self.checksum = [{"Checksum": "Sum"}]
 
-    @mock.patch("mysql_db_admin.gen_libs.prt_msg")
-    @mock.patch("mysql_db_admin.mysql_libs.analyze_tbl")
-    def test_run_analyze(self, mock_analyze, mock_prt):
+    @mock.patch("mysql_db_admin.mysql_libs.checksum")
+    def test_run_checksum(self, mock_checksum):
 
-        """Function:  test_run_analyze
+        """Function:  test_run_checksum
 
-        Description:  Test run_analyze function.
+        Description:  Test run_checksum function.
 
         Arguments:
 
         """
 
-        mock_analyze.return_value = self.analyze_tables
-        mock_prt.return_value = True
+        mock_checksum.return_value = self.checksum
 
         with gen_libs.no_std_out():
-            self.assertFalse(mysql_db_admin.run_analyze(self.server, "db",
-                                                        "tbl"))
+            self.assertFalse(mysql_db_admin.run_checksum(self.server, "db",
+                                                         "tbl"))
 
 
 if __name__ == "__main__":